pyproject.toml
[project]
name = "catalog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arrow>=1.3.0",
    "bs4>=0.0.2",
    "click>=8.2.1",
    "fastapi[all]>=0.116.1",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.28",
    "langchain-text-splitters>=0.3.8",
    "lxml>=6.0.0",
    "ollama>=0.5.1",
    "openai>=1.97.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "sentence-transformers>=5.0.0",
]

[dependency-groups]
dev = [
    "rich>=14.0.0",
    "ruff>=0.12.2",
]
